Basic deferred shading

See merge request sergcpp/DummyEngine!18

+++ src/Eng/Renderer/Passes/RpRTReflections.h
@@ -17,7 +17,7 @@ class RpRTReflections : public RenderPassBase {
     Ren::Pipeline pi_rt_reflections_inline_;
 
     // temp data (valid only between Setup and Execute calls)
-    uint32_t render_flags_ = 0;
+    uint64_t render_flags_ = 0;
     const ViewState *view_state_ = nullptr;
     const Ren::Camera *draw_cam_ = nullptr;
     const AccelerationStructureData *acc_struct_data_ = nullptr;
